clock/v2: Rename `Tokens::new` to `retrieve_tokens`; flatten `sources` module
