+++ hal/src/thumbv7em/clock/v2/gclkio.rs
@@ -5,13 +5,13 @@ use core::marker::PhantomData;
 use seq_macro::seq;
 use typenum::U0;
 
-use crate::clock::types::{Enabled, Counter, Decrement, Increment};
+use crate::clock::types::{Counter, Decrement, Enabled, Increment};
 use crate::clock::v2::{Source, SourceMarker};
 use crate::gpio::v2::{self as gpio, AlternateM, AnyPin, Pin, PinId};
 use crate::time::Hertz;
 use crate::typelevel::*;
 
-use super::super::gclk::*;
+use super::gclk::*;
 
 //==============================================================================
 // GclkIo
