+++ hal/src/thumbv7em/clock/v2/pclk.rs
@@ -9,14 +9,14 @@ use crate::pac;
 
 pub use crate::pac::gclk::pchctrl::GEN_A as PclkSourceEnum;
 
-use crate::clock::types::{Enabled, Counter, Decrement, Increment};
+use crate::clock::types::{Counter, Decrement, Enabled, Increment};
 use crate::clock::v2::{Source, SourceMarker};
 use crate::sercom::v2::*;
 use crate::time::Hertz;
 use crate::typelevel::Sealed;
 
+use super::dpll::{Pll0, Pll1};
 use super::gclk::*;
-use super::sources::dpll::{Pll0, Pll1};
 
 //==============================================================================
 // PclkToken
